Move AccountSelectionSpinner into its own file
