+++ app/ui/src/main/res/layout/account_spinner_dropdown_item.xml
@@ -5,6 +5,9 @@
     android:layout_height="match_parent"
     android:gravity="center_vertical"
     android:orientation="vertical"
+    android:paddingTop="8dp"
+    android:paddingBottom="8dp"
+    android:paddingLeft="16dp"
     android:paddingRight="16dp">
 
     <TextView
